chore: Move context menu to be closer to annotator context menu
