+++ web_ui/src/pages/annotator/providers/annotator-context-menu-provider/annotator-context-menu-provider.component.tsx
@@ -6,9 +6,9 @@ import { createContext, ReactNode, useContext, useReducer, useState } from 'reac
 import isEqual from 'lodash/isEqual';
 import noop from 'lodash/noop';
 
-import { ContextMenu, ContextMenuProps } from '../../../../shared/components/context-menu/context-menu.component';
 import { MissingProviderError } from '../../../../shared/missing-provider-error';
 import { useZoom } from '../../zoom/zoom-provider.component';
+import { ContextMenu, ContextMenuProps } from './context-menu/context-menu.component';
 import { getContextMenuPosition } from './utils';
 
 interface AnnotatorContextConfig
