+++ web_ui/src/pages/annotator/providers/annotator-context-menu-provider/context-menu/context-menu.module.scss
